# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/expenses:
    get:
      tags:
        - Expenses
      summary: List Expenses
      description: Get a paginated list of Expenses with optional filtering and sorting
      operationId: listExpenses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Expenses
      summary: Create Expense
      description: Create a new Expense
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/expense_categories:
    get:
      tags:
        - ExpenseCategories
      summary: List ExpenseCategories
      description: Get a paginated list of ExpenseCategories with optional filtering and sorting
      operationId: listExpenseCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ExpenseCategories
      summary: Create ExpenseCategory
      description: Create a new ExpenseCategory
      operationId: createExpenseCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Expense:
      type: object
      required:
        - id
        - employee_id
        - category_id
        - expense_date
        - description
        - amount_total
        - currency
        - payment_mode
        - approval_state
        - state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        expense_date:
          type: string
          format: date
        description:
          type: string
        amount_total:
          type: string
        currency:
          type: string
          maxLength: 3
        payment_mode:
          type: string
        reference:
          nullable: true
          type: string
        approval_state:
          type: string
        state:
          type: string
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        reimbursement_id:
          nullable: true
          type: string
          format: uuid
        receipt_file_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateExpenseRequest:
      type: object
      required:
        - employee_id
        - category_id
        - expense_date
        - description
        - amount_total
        - currency
        - payment_mode
        - approval_state
        - state
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        expense_date:
          type: string
          format: date
        description:
          type: string
        amount_total:
          type: string
        currency:
          type: string
          maxLength: 3
        payment_mode:
          type: string
        reference:
          nullable: true
          type: string
        approval_state:
          type: string
        state:
          type: string
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        reimbursement_id:
          nullable: true
          type: string
          format: uuid
        receipt_file_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    ExpenseList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ExpenseCategory:
      type: object
      required:
        - id
        - code
        - name
        - expense_account_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        expense_account_id:
          type: string
          format: uuid
        default_tax_rate:
          nullable: true
          type: string
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateExpenseCategoryRequest:
      type: object
      required:
        - code
        - name
        - expense_account_id
        - metadata
      properties:
        code:
          type: string
        name:
          type: string
        expense_account_id:
          type: string
          format: uuid
        default_tax_rate:
          nullable: true
          type: string
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    ExpenseCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []