# src/exports/types.rs
//! Public types for Expenses module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate};
use rust_decimal::Decimal;
use crate::domain::entity::*;

// ============================================================================
// EXPENSE TYPES
// ============================================================================

/// Type-safe ID for Expense
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ExpenseId(pub Uuid);

impl ExpenseId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ExpenseId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ExpenseId> for Uuid {
    fn from(id: ExpenseId) -> Self {
        id.0
    }
}

/// Data transfer object for Expense
///
/// This is the public representation of Expense for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseDto {
    pub id: ExpenseId,
    pub employee_id: Uuid,
    pub category_id: Uuid,
    pub expense_date: NaiveDate,
    pub description: String,
    pub amount_total: Decimal,
    pub currency: String,
    pub payment_mode: ExpensePaymentMode,
    pub reference: Option<String>,
    pub approval_state: ExpenseApprovalState,
    pub state: ExpenseState,
    pub approval_request_id: Option<Uuid>,
    pub journal_id: Option<Uuid>,
    pub accounting_post_id: Option<Uuid>,
    pub reimbursement_id: Option<Uuid>,
    pub receipt_file_id: Option<Uuid>,
    pub metadata: serde_json::Value,
}

/// Summary view of Expense for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseSummary {
    pub id: ExpenseId,
}

/// Reference to Expense for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseRef {
    pub id: ExpenseId,
}

// ============================================================================
// EXPENSECATEGORY TYPES
// ============================================================================

/// Type-safe ID for ExpenseCategory
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ExpenseCategoryId(pub Uuid);

impl ExpenseCategoryId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ExpenseCategoryId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ExpenseCategoryId> for Uuid {
    fn from(id: ExpenseCategoryId) -> Self {
        id.0
    }
}

/// Data transfer object for ExpenseCategory
///
/// This is the public representation of ExpenseCategory for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseCategoryDto {
    pub id: ExpenseCategoryId,
    pub code: String,
    pub name: String,
    pub expense_account_id: Uuid,
    pub default_tax_rate: Option<Decimal>,
    pub description: Option<String>,
    pub metadata: serde_json::Value,
}

/// Summary view of ExpenseCategory for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseCategorySummary {
    pub id: ExpenseCategoryId,
    pub name: String,
}

/// Reference to ExpenseCategory for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseCategoryRef {
    pub id: ExpenseCategoryId,
}

// ============================================================================
// EXPENSETAXLINE TYPES
// ============================================================================

/// Type-safe ID for ExpenseTaxLine
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct ExpenseTaxLineId(pub Uuid);

impl ExpenseTaxLineId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for ExpenseTaxLineId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ExpenseTaxLineId> for Uuid {
    fn from(id: ExpenseTaxLineId) -> Self {
        id.0
    }
}

/// Data transfer object for ExpenseTaxLine
///
/// This is the public representation of ExpenseTaxLine for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseTaxLineDto {
    pub id: ExpenseTaxLineId,
    pub expense_id: Uuid,
    pub basis: String,
    pub account_id: Uuid,
    pub description: Option<String>,
    pub rate: Decimal,
    pub tax_amount: Decimal,
    pub metadata: serde_json::Value,
}

/// Summary view of ExpenseTaxLine for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseTaxLineSummary {
    pub id: ExpenseTaxLineId,
}

/// Reference to ExpenseTaxLine for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ExpenseTaxLineRef {
    pub id: ExpenseTaxLineId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>